jukebox: move ciso code to a file, improve comments, rename some symbols
