+++ Externals/SlippiRustExtensions/jukebox/src/scenes.rs
@@ -1,4 +1,4 @@
-/// Sourced from M'OVerlay: https://github.com/bkacjios/m-overlay/blob/d8c629d/source/melee.lua
+/// Sourced from M'Overlay: https://github.com/bkacjios/m-overlay/blob/d8c629d/source/melee.lua
 #[rustfmt::skip]
 #[allow(dead_code)]
 pub(crate) mod scene_ids {
